feat. MainPage API 연동 + 리펙토링 todo. statePage API 연동 + 리펙토링 #42
