+++ src/apis/mainPage.ts
@@ -1,0 +1,9 @@
+import api from './common';
+
+export const mainApi = {
+  getBefore: async () => api.get('main/before'),
+
+  getComplete: async () => api.get('main/complete'),
+
+  getRequesting: async () => api.get('main/requesting'),
+};
