+++ src/components/CardList.tsx
@@ -1,18 +1,15 @@
 import { Card } from '../components/index';
-import { IDummyList } from 'src/dummy/dummyList';
 import styled from 'styled-components';
-import { IData } from './MainRequestBoard';
+import { IData } from '../containers/mainPage/components/MainRequestBoard';
 
 interface CardListType {
-  offset: number;
-  limit: number;
   list: IData[];
 }
 
-export const CardList = ({ list, offset, limit }: CardListType) => {
+export const CardList = ({ list }: CardListType) => {
   return (
     <CardBox>
-      {list.slice(offset, offset + limit).map(cardInfo => (
+      {list.map(cardInfo => (
         <Card key={cardInfo.id} {...cardInfo} />
       ))}
     </CardBox>
