+++ src/containers/mainPage/MainPage.tsx
@@ -1,15 +1,11 @@
 import type { NextPage } from 'next';
-import { Typography, SideBox, MainRequestBoard } from '../../components/index';
+import { SideBox, MainRequestBoard } from '../../components/index';
 import styled from 'styled-components';
-import Link from 'next/link';
-import { all } from 'src/constants/all/all';
-import { pathName } from 'src/constants/pathName';
 import { useRouter } from 'next/router';
-import axios from 'axios';
-import { useEffect, useState } from 'react';
+import Comment from './components/Comment';
+import Write from './components/Write';
 
 export const MainPage: NextPage = () => {
-  const [data, setData] = useState({});
   const router = useRouter();
   const callKaKaoLoginHandler = () => {
     router.push({
@@ -22,28 +18,15 @@ export const MainPage: NextPage = () => {
     });
   };
 
-  const onClick = () => {
-  };
-
   return (
     <Container>
       <button onClick={() => callKaKaoLoginHandler()}>
         카카오 로그인 테스트
       </button>
-      <SignUpBox>
-        <SignUpComment onClick={() => onClick()}>
-          <Typography size="40" color="blue" fontWeight="900">
-            {all.comment}
-          </Typography>
-        </SignUpComment>
-        <Link href={pathName.SIGNUP} passHref>
-          <SignUpBtn>
-            <Typography size="20" color="white" fontWeight="900">
-              {all.writeBtn}
-            </Typography>
-          </SignUpBtn>
-        </Link>
-      </SignUpBox>
+      <WriteBox>
+        <Comment />
+        <Write />
+      </WriteBox>
 
       <MainRequestBoard state="before" />
       <MainRequestBoard state="doing" />
@@ -64,27 +47,10 @@ const Container = styled.div`
   flex-direction: column;
 `;
 
-const SignUpBox = styled.div`
+const WriteBox = styled.div`
   width: 100%;
   height: 140px;
   display: flex;
   flex-direction: column;
   margin-bottom: 150px;
-`;
-
-const SignUpComment = styled.div`
-  text-align: center;
-`;
-
-const SignUpBtn = styled.button`
-  margin: 60px auto 0 auto;
-  width: 185px;
-  background-color: rgba(7, 104, 159, 1);
-  border-radius: 10px;
-  padding: 10px 27px;
-  border: none;
-  :hover {
-    background-color: black;
-    cursor: pointer;
-  }
 `;
