+++ src/containers/mainPage/StatePage.tsx
@@ -9,20 +9,17 @@ import { dummyFilter } from 'src/constants/all/filter';
 import { dummyList } from 'src/dummy/dummyList';
 import { all } from 'src/constants/all/all';
 import { BiSearchAlt2 } from 'react-icons/bi';
-import React, { useMemo, useState } from 'react';
+import React, { useState } from 'react';
 import { useRouter } from 'next/router';
 import Link from 'next/link';
-import axios from 'axios';
 import { pathName } from 'src/constants/pathName';
 
 const StatePage = () => {
   const [text, setText] = useState(''); // 필터링 값
   const [selected, setSelected] = useState('title');
   const [page, setPage] = useState(1);
-  const offset = useMemo(() => (page - 1) * 16, [page]);
   const router = useRouter();
   const { state } = router.query;
-  const [post, setPost] = useState([]); // 데이터 받아와서 저장하는 state
 
   const handlePageChange = (page: number) => {
     setPage(page);
@@ -81,7 +78,7 @@ const StatePage = () => {
         <Hr />
       </div>
       <CardListWrap>
-        <CardList list={dummyList} offset={offset} limit={16} />
+        <CardList list={dummyList} />
       </CardListWrap>
       <CustomPagination
         itemClass="page"
@@ -206,6 +203,5 @@ const Hr = styled.hr`
   margin-bottom: 20px;
 `;
 const A = styled.a`
-  text-decoration: none;
   text-align: center;
 `;
