+++ styles/globals.css
@@ -164,3 +164,7 @@ body {
   right: 0;
   bottom: -25px;
 }
+
+a {
+  text-decoration: none;
+}
